Added ability to add images to the default image library media package from the YUI HTML editor.

+++ include/media.model.php
@@ -898,12 +898,18 @@
 			if($uploadingFile)
 			{
 				$oldPath = $uploadfolder."/".$mediaFile->get_media_file_id().".*";
-				foreach(glob($oldPath) as $oldFile)
-				{
-					// echo($oldFile."<br>");
-					unlink($oldFile);
-				}
 
+				$files = glob($oldPath);
+				// print_r($files);
+
+				if(is_array($files) && count($files) > 0)
+				{
+					foreach($files as $oldFile)
+					{
+						// echo($oldFile."<br>");
+						unlink($oldFile);
+					}
+				}
 				// die();
 
 				$newPath = $uploadfolder."/".$mediaFile->get_media_file_id().".".$extension;
